Separate out base class into its own file
